Responsividade e README finalizados.
