+++ pokedex/src/components/Header/header-style.js
@@ -1,0 +1,28 @@
+import styled from "styled-components";
+
+export const Button = styled.button`
+  margin-left: 6.5%;
+  padding: 10px;
+  border: none;
+  border-radius: 6px;
+  background-color: #596d82;
+  color: #ffffff;
+  cursor: pointer;
+  font-size: 0.8em;
+  font-family: joystix;
+  :hover {
+    background-color: #85a1ca;
+  }
+`;
+
+export const Title = styled.h1`
+  color: #d29a95;
+  font-size: 1.4em;
+  margin: 0 auto;
+  margin: 0 15px 0 15px;
+`;
+
+export const Pokebola = styled.img`
+  max-width: 35px;
+  margin-right: 90px;
+`;
