+++ pokedex/src/components/PokeCard/PokeCard.js
@@ -1,63 +1,17 @@
 import React, { useContext } from "react";
-import { useHistory } from "react-router-dom";
 import GlobalStateContext from "../../global/GlobalStateContext";
+import { useHistory } from "react-router-dom";
 import { goToDetailsPage } from "../../routes/coordinator";
 import { Box } from "@chakra-ui/react";
 import { Image } from "@chakra-ui/react";
-import styled from "styled-components";
+import { Flex } from "@chakra-ui/react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faEye,
   faMinusCircle,
   faPlusCircle,
 } from "@fortawesome/free-solid-svg-icons";
-import { Flex } from "@chakra-ui/react";
-import { createBreakpoints } from "@chakra-ui/theme-tools";
-
-const breakpoints = createBreakpoints({
-  sm: "30em",
-  md: "48em",
-  lg: "62em",
-  xl: "80em",
-  "2xl": "96em",
-});
-
-const PokeName = styled.p`
-  color: #ffffff;
-  text-shadow: 2px 2px 2.5px gray;
-  font-family: joystix;
-  margin: 0 auto;
-  margin-top: 13px;
-  margin-bottom: 8px;
-  text-align: center;
-  font-size: small;
-`;
-
-const AddAndRemoveButton = styled.button`
-  width: 60px;
-  background-color: transparent;
-  border: none;
-  color: #008000;
-  cursor: pointer;
-  :hover {
-    color: #ffffff;
-    transform: scale(1.2);
-    transition: all 0.3s ease 0s;
-  }
-`;
-
-const DetailsButton = styled.button`
-  width: 60px;
-  background-color: transparent;
-  border: none;
-  color: #4869a2;
-  cursor: pointer;
-  :hover {
-    color: #ffffff;
-    transform: scale(1.2);
-    transition: all 0.3s ease 0s;
-  }
-`;
+import { PokeName, AddAndRemoveButton, DetailsButton } from "./pokecard-style";
 
 const PokeCard = (props) => {
   const history = useHistory();
