+++ pokedex/src/pages/PokedexPage.js
@@ -1,26 +1,10 @@
 import React, { useContext } from "react";
+import GlobalStateContext from "../global/GlobalStateContext";
+import Header from "../components/Header/Header";
+import PokeCard from "../components/PokeCard/PokeCard";
 import { useHistory } from "react-router-dom";
 import { goToLastPage } from "../routes/coordinator";
-import Header from "../components/Header/Header";
-import styled from "styled-components";
-import GlobalStateContext from "../global/GlobalStateContext";
-import PokeCard from "../components/PokeCard/PokeCard";
 import { Flex } from "@chakra-ui/react";
-
-const Button = styled.button`
-  width: 150px;
-  padding: 10px;
-  border: none;
-  border-radius: 6px;
-  background-color: #596d82;
-  color: #ffffff;
-  cursor: pointer;
-  :hover {
-    background-color: #85a1ca;
-  }
-  display: block;
-  margin: 0 auto;
-`;
 
 const PokedexPage = () => {
   const history = useHistory();
