+++ pokedex/src/pages/detailspage/detailspage-style.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Lists = styled.li`
+  list-style: none;
+`;
+
+export const Titles = styled.p`
+  margin-left: 66px;
+  color: #ffffff;
+  text-shadow: 2px 2px 2.5px gray;
+`;
